also checks outputs of parent calculation for retrieving structure and kpoints
The projwfc parser implicitly retreives structure and kpoints from
the inputs of parent calculation of remote_folder, this works if
the parent calculation is a PwCalculation. However, if the projwfc
calculation is performed after an OpengridCalculation, it will
fails since OpengridCalculation outputs a newly unfolded kpoints
mesh. Now projwfc parser can use this output kpoints mesh to do
the parsing.