`PdosWorkChain`: Fix initalization and protocol
Fixes two issues when using the `PdosWorkChain`:

1. Remove the validation of the `parent_folder` input for `nscf`
`PwCalculation`s. This made it impossible to use the `PdosWorkChain`
with an initial `scf` run, since the `parent_folder` is only created
inside the logic of the work chain, and hence cannot be provided as an
input upon submission.
2. In the `PdosWorkChain.get_builder_from_protocol()` method, override
the `occupations` setting to make sure it is set to `tetrahedra`.
Although this setting is also passed to the overrides from the `yaml`
file that defines the protocol, if the user sets
`ElectronicType.INSULATOR`, the occupations is set to `fixed` in the
`PwBaseWorkChain.get_builder_from_protocol()` method.